+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+-   Adds bindings for `ctrl+shift+home` and `ctrl+shift+end` to select while moving to document start/end.
+
 ## [0.12.0] - 2024-04-17
 
 -   Show the computed filepath in the Save and Open widgets ([#232](https://github.com/tconbeer/textual-textarea/issues/232) - thank you [@bjornasm](https://github.com/bjornasm)!).